move customprovider to shared package and increase consistency with other providers
