+++ packages/shared/package.json
@@ -53,6 +53,7 @@
     "kysely": "^0.27.3",
     "langchain": "^0.1.36",
     "lodash": "^4.17.21",
+    "next-auth": "^4.24.7",
     "prisma-extension-kysely": "^2.1.0",
     "zod": "^3.22.4",
     "zod-to-json-schema": "^3.22.5"
